+++ alice/helper/github_helper.py
@@ -1,7 +1,8 @@
 import json
 import requests
 from alice.helper.api_manager import ApiManager
-from alice.helper.constants  import API_GITHUB_REVIEW_ACCEPT_KEY, EP_REVIEWS, API_GITHUB_REPO_MEMBER
+from alice.helper.constants  import API_GITHUB_REVIEW_ACCEPT_KEY, EP_REVIEWS, API_GITHUB_REPO_MEMBER, API_GITHUB_ISSUES, \
+    EP_COMMENTS
 from alice.helper.decorators.retry import Retry
 from alice.helper.log_utils import LOG
 
@@ -12,14 +13,22 @@ class PRFilesNotFoundException(Exception):
         super(PRFilesNotFoundException, self).__init__(str(self.pr_response))
 
 
+class PRContentNotFoundException(Exception):
+    def __init__(self, pr_response):
+        self.pr_response = pr_response
+        super(PRContentNotFoundException, self).__init__(str(self.pr_response))
+
+
+
 class GithubHelper(object):
 
-    def __init__(self, org, repo, github_token, pr_api_link):
-        self.pr_api_link = pr_api_link
-        self.GITHUB_TOKEN = github_token
+    def __init__(self, pr):
+        self.GITHUB_TOKEN = pr.config.githubToken
+        self.pr_api_link =  pr.link
         self.headers = {"Authorization": "token " + self.GITHUB_TOKEN}
+        self.pr = pr
 
-        url = API_GITHUB_REPO_MEMBER.format(org=org, repo=repo)
+        url = API_GITHUB_REPO_MEMBER.format(org=self.pr.config.organisation, repo=self.pr.repo)
         response = ApiManager.get(url=url, headers=self.headers)
         if response["status_code"] != 200:
             raise Exception(response["content"], "Please check the provided Github Token, "
@@ -59,4 +68,9 @@ class GithubHelper(object):
             raise PRFilesNotFoundException(files_content)
         return files_content
 
+    @Retry(PRContentNotFoundException, max_retries=3)
+    def get_comments(self):
+        comments_end_point = API_GITHUB_ISSUES.format(org=self.pr.config.organisation, repo=self.pr.repo) \
+                             + "/" + str(self.pr.number) + "/" + EP_COMMENTS
+        return ApiManager.get(comments_end_point, self.headers)
