+++ docs/samples/config.json
@@ -1,106 +1,113 @@
 {
-    "organisation": "moengage",
-    "tokens":{
-        "github": "",
-        "slack": "",
-        "jenkins": ""
-    },
-    "repo" :{
-      "MoEngage": {
-        "sensitive_branches": [
-          "dev",
-          "qa",
-          "master"
-        ],
-        "sensitive_files": [""],
-        "sensitive_file_pattern": "",
-        "product_plus_required_dir_pattern": "dashboard",
-        "cc_members": "",
-        "code_channel": "#code-merged",
-        "alert_channel": "#weekly-releases",
-        "product_team": "",
-        "dev_ops_team": "",
-        "notify_direct" : {
-          "action_to_be_notified_on": "opened",
-          "branch_list_to_be_notified": ["dev", "qa"],
-          "person_to_be_notified": "",
-          "tech_leads_to_be_notified_on_release_freeze": []
-        },
-        "main_branch" : "master",
-        "test_branch" : "qa",
-        "whitelisted_git_members": [
-          "shahp00ja",
-          "satyamoengage",
-          "snehsagar",
-          "vandanamoriwal",
-          "geetima12"
-        ],
-        "super_git_members":[
-          "shahp00ja",
-          "satyamoengage"
-        ],
-        "checks": [ "close_dangerous_pr", "tech_review", "product_review", "comment_guidelines",
-          "notify_lead_on_given_action", "notify_channel_on_merge", "notify_on_sensitive_files_touched",
-          "notify_qa_sign_off", "remind_direct_release_guideline_on_merge"]
-      },
-      "segmentation": {
-        "sensitive_branches": [
-          "develop",
-          "release",
-          "qa",
-          "master"
-        ],
-        "cc_members": "@akshay @pooja",
-        "code_channel": "#segmentation-code",
-        "alert_channel": "#weekly-releases",
-        "product_team": "@riteshg @nalin",
-        "dev_ops_team": "@pruthvi @adithya",
-        "lead_to_be_notified": "@yashwanth",
-        "whitelisted_git_members": [
-          "akgoel-mo",
-          "shahp00ja"
-        ]
-      },
-      "commons": {
-        "sensitive_branches": [
-          "develop",
-          "release",
-          "qa",
-          "master"
-        ],
-        "cc_members": "@akshay @mshekher @pooja",
-        "code_channel": "#commons-code",
-        "alert_channel": "#weekly-releases",
-        "product_team": "@riteshg @nalin",
-        "dev_ops_team": "@pruthvi @adithya",
-        "lead_to_be_notified": "@yashwanth",
-        "whitelisted_git_members": [
-          "akgoel-mo",
-          "shahp00ja",
-          "pruthvintss"
-        ]
-      }
-    },
-
-  "dev_ops_team": "@pruthvi @adithya",
-  "lead_to_be_notified": "@yashwanth",
-  "JENKINS_BASE": "http://ci.moengage.com/",
-  "repo_list": ["MoEngage", "MoeDataScience", "mongoengine", "inapp-rest-service", "mongo", "segmentation",
-                "commons", "moengcache", "MoeAbTesting", "apns_client", "Bugsy_Backend", "product-management",
-                "product-recommendation"],
-  "debug": true,
-  "debug_alice": {
-		"sensitive_branches": [],
-        "debug_channel": "#tmp",
-        "debug_folks" : "@pooja",
-        "whitelisted_git_members": ["shahp00ja"]
-  },
-  "user_map":{
-    "shahp00ja":"pooja",
-    "abhi9moe":"abhinav",
-    "moarnab":"arnab",
-    "geetima12":"geetima",
-    "naveenkumarkokku":"naveenkumarkokku"
-  },
-  "release_notes_link": "https://docs.google.com/a/moengage.com/spreadsheets/d/1eW3y-GxGzu8Jde8z4EYC6fRh1Ve4TpbW5qv2-iWs1ks/edit?usp=sharing"
+	"organisation": "",
+	"tokens": {
+		"github": "",
+		"slack": "",
+		"jenkins": ""
+	},
+	"dev_ops_team": "@p1 @p2",
+	"JENKINS_BASE": "http://YOUR_JENKINS_DOMAIN.com/",
+	"repo_list": [
+		"REPO_1",
+		"REPO_2"
+	],
+	"debug": true,
+	"debug_alice": {
+		"debug_channel": "#tmp",
+		"debug_folks": "@pooja"
+	},
+	"user_map": {
+		"p00ja": "pooja",
+		"abhi": "abhinav"
+	},
+	"release_notes_link": "",
+	"repo": {
+		"REPO_1": {
+			"checks": [
+				"close_dangerous_pr",
+				"tech_review",
+				"product_review",
+				"comment_guidelines",
+				"notify_lead_on_given_action",
+				"notify_channel_on_merge",
+				"notify_on_sensitive_files_touched",
+				"notify_qa_sign_off",
+				"remind_direct_release_guideline_on_merge"
+			],
+			"sensitive_branches": [
+				"dev",
+				"qa",
+				"master"
+			],
+			"sensitive_files": [
+				"f1",
+				"f2"
+			],
+			"product_team": "@p1 @p2 @p3",
+			"product_plus_required_dir_pattern": "dashboard",
+			"product_team_github_names": [
+				"p1",
+				"p2",
+				"p3"
+			],
+			"cc_members": "@p1 @p2",
+			"code_channel": "#repo1-code",
+			"alert_channel": "#weekly-releases",
+			"notify_direct": {
+				"person_to_be_notified": "@yashwanth",
+				"action_to_be_notified_on": "opened",
+				"branch_list_to_be_notified": [
+					"dev",
+					"qa"
+				],
+				"tech_leads_to_be_notified_on_release_freeze": [
+					"@p1",
+					"@p2"
+				]
+			},
+			"main_branch": "master",
+			"test_branch": "qa",
+			"super_git_members": [
+				"p1",
+				"p2"
+			]
+		},
+		"REPO_2": {
+			"checks": [
+				"tech_review"
+			],
+			"sensitive_branches": [
+				"dev",
+				"qa",
+				"master"
+			],
+			"sensitive_files": [
+				"f1",
+				"f2"
+			],
+			"product_team": "@p1",
+			"product_plus_required_dir_pattern": "dashboard",
+			"cc_members": "@p1 @p2 @p3",
+			"code_channel": "#repo2-code",
+			"alert_channel": "#weekly-releases",
+			"notify_direct": {
+				"person_to_be_notified": "@p",
+				"action_to_be_notified_on": "opened",
+				"branch_list_to_be_notified": [
+					"dev",
+					"qa"
+				],
+				"tech_leads_to_be_notified_on_release_freeze": [
+					"@p1",
+					"@p2"
+				]
+			},
+			"main_branch": "master",
+			"test_branch": "qa",
+			"super_git_members": [
+				"p1",
+				"p2"
+			]
+		}
+	}
 }
